Reverted changes in genreco.py
Note: Setting a value to num_gpu will likely make tofu ez crash because of the mismatch in indices (e.g. setting "1" will try to access gpu[1] even if there is one GPU in computer (i.e. only gpu[0] exists)). This is a known problem. However, the change was reverted because modifying genreco.py was outside the scope of this project.