Removed 'slices-per-device' setting from Tofu-ez
From a discussion with Sergey, we determined that a user will only want to use 'data-splitting-policy' and 'slice-memory-coeff', and additional settings will likely confuse the user. If a user wants additional control, they can use the ufo-kit commands instead. In addition, setting a value to 'slices-per-device' or 'num-of-gpus' will always cause the reconstruction to stop. Thus, 'slices-per-device' was removed from Tofu-ez GUI.

Also, the check for an empty strings are now changed to "None" value instead. The "None" values are better handled for data types that don't handle string inputs.